+++ pcs/daemon/ssl.py
@@ -30,20 +30,27 @@ def check_cert_key(cert_path, key_path):
     if errors:
         return errors
 
-    context = SSL.Context(SSL.TLSv1_METHOD)
-    context.use_privatekey(key)
-    context.use_certificate(cert)
+    try:
+        context = SSL.Context(SSL.TLSv1_METHOD)
+        context.use_privatekey(key)
+        context.use_certificate(cert)
+    except SSL.Error as e:
+        errors.append(f"Unable to load SSL certificate and/or key: {e}")
+        # If we cannot load the files, do not confuse users with other error
+        # messages.
+        return errors
     try:
         context.check_privatekey()
-    except (crypto.Error, SSL.Error):
-        errors.append("SSL certificate does not match the key")
+    except (crypto.Error, SSL.Error) as e:
+        errors.append(f"SSL certificate does not match the key: {e}")
+
     return errors
 
 def open_ssl_file_to_rewrite(path):
     return os.fdopen(os.open(path, os.O_CREAT|os.O_WRONLY, 0o600), "wb")
 
-def regenerate_cert_key(server_name, cert_path, key_path):
-    key = generate_key()
+def regenerate_cert_key(server_name, cert_path, key_path, key_length=None):
+    key = generate_key(key_length) if key_length else generate_key()
     with open_ssl_file_to_rewrite(cert_path) as cert_file:
         cert_file.write(dump_cert(generate_cert(key, server_name)))
     with open_ssl_file_to_rewrite(key_path) as key_file:
@@ -72,8 +79,13 @@ class CertKeyPair:
     def check(self):
         return check_cert_key(self.cert_location, self.key_location)
 
-    def regenerate(self, server_name):
-        regenerate_cert_key(server_name, self.cert_location, self.key_location)
+    def regenerate(self, server_name, key_length=None):
+        regenerate_cert_key(
+            server_name,
+            self.cert_location,
+            self.key_location,
+            key_length
+        )
 
 class SSLCertKeyException(Exception):
     pass